Changelog:

- Reformatted `crystal17.main` and `crystal17.basic` to use run program directly from `crystal` executable (rather than the `runcry` script)
  - added restart functionality to `CryMainCalculation` (*via* a fort.9 in a remote folder)
  - added checks and error codes for PbsPro messages to `_scheduler_stderr.txt` (e.g. walltime limit reached)
  - allow SHRINK IS input to be a list ([IS1, IS2, IS3])
  - added output of `TrajectoryData` of optimisation steps for `CryMainCalculation`
- Wrote internal stdout parser for `crystal17.main`
  - This builds and improves on the original ejplugins implementation: making the parsing flow more easy to understand, adding additional data parsing (some taken from tilde), and restructuring the output json
- added `CryMainBaseWorkChain` (a replica of `PwBaseWorkChain`  from `aiida-quanumespresso`)
- Added `crystal17.doss` and `crystal.fermi` calculations, and `CryPropertiesWorkChain`
  - also added add band gap calcfunction
- Improved `crystal17.sym3d`
- Standardised GULP potentials and rewrote `PotentialData`
  - all potentials now share the a common jsonschema
  - the full potential file is created and stored in the repository
  - refactored reaxff read/write
  - added fitting data input
- Added `GulpFittingCalculation`
- Added some helpful methods for manipulating StructureData
- Rewrote immigration functions (to use `aiida-core v1` mechanics)
- Restructure tests into central location and added many more
- Updates Documentation and removed `graph.py` (now in aiida-core)

+++ conftest.py
@@ -9,7 +9,7 @@ from aiida.manage.fixtures import fixture_manager
 import pytest
 
 from aiida_crystal17.tests.utils import AiidaTestApp
-from aiida_crystal17.tests import TEST_DIR
+from aiida_crystal17.tests import TEST_FILES, get_test_structure, get_test_structure_and_symm
 
 
 @pytest.fixture(scope='session')
@@ -25,69 +25,49 @@ def aiida_environment():
 @pytest.fixture(scope='function')
 def db_test_app(aiida_environment):
     """clear the database after each test"""
-
-    if os.environ.get("MOCK_CRY17_EXECUTABLES", False):
+    if os.environ.get("MOCK_CRY17_EXECUTABLES", True) != "false":
         print("NB: using mock executable")
         executables = {
-            'crystal17.basic': 'mock_runcry17',
-            'crystal17.main': 'mock_runcry17',
+            'crystal17.basic': 'mock_crystal17',
+            'crystal17.main': 'mock_crystal17',
+            'crystal17.doss': 'mock_properties17',
+            'crystal17.fermi': 'mock_properties17',
             'gulp.single': 'mock_gulp',
-            'gulp.optimize': 'mock_gulp'
+            'gulp.optimize': 'mock_gulp',
+            'gulp.fitting': 'mock_gulp'
         }
     else:
         executables = {
-            'crystal17.basic': 'runcry17',
-            'crystal17.main': 'runcry17',
+            'crystal17.basic': 'crystal17',
+            'crystal17.main': 'crystal17',
+            'crystal17.doss': 'properties17',
+            'crystal17.fermi': 'properties17',
             'gulp.single': 'gulp',
-            'gulp.optimize': 'gulp'
+            'gulp.optimize': 'gulp',
+            'gulp.fitting': 'gulp'
         }
 
-    work_directory = tempfile.mkdtemp()
+    test_workdir = os.environ.get("CRY17_TEST_WORKDIR", None)
+    if test_workdir:
+        print("NB: using test workdir: {}".format(test_workdir))
+        work_directory = test_workdir
+    else:
+        work_directory = tempfile.mkdtemp()
     yield AiidaTestApp(
         work_directory, executables, environment=aiida_environment)
     aiida_environment.reset_db()
-    shutil.rmtree(work_directory)
+    if not test_workdir:
+        shutil.rmtree(work_directory)
 
 
 @pytest.fixture(scope='function')
 def get_structure():
-    def _get_structure(name):
-        from aiida.plugins import DataFactory
-        from ase.spacegroup import crystal
-        structure_data_cls = DataFactory('structure')
-        if name == "MgO":
-            atoms = crystal(
-                symbols=[12, 8],
-                basis=[[0, 0, 0], [0.5, 0.5, 0.5]],
-                spacegroup=225,
-                cellpar=[4.21, 4.21, 4.21, 90, 90, 90])
-            return structure_data_cls(ase=atoms)
-        elif name == "NiO_afm":
-            atoms = crystal(
-                symbols=[28, 8],
-                basis=[[0, 0, 0], [0.5, 0.5, 0.5]],
-                spacegroup=225,
-                cellpar=[4.164, 4.164, 4.164, 90, 90, 90])
-            atoms.set_tags([1, 1, 2, 2, 0, 0, 0, 0])
-            return structure_data_cls(ase=atoms)
-        elif name == "pyrite":
-            from aiida_crystal17.symmetry import convert_structure
-            structure_data = {
-                "lattice": [[5.38, 0.000000, 0.000000],
-                            [0.000000, 5.38, 0.000000],
-                            [0.000000, 0.000000, 5.38]],
-                "fcoords": [[0.0, 0.0, 0.0], [0.5, 0.0, 0.5], [0.0, 0.5, 0.5],
-                            [0.5, 0.5, 0.0], [0.338, 0.338, 0.338],
-                            [0.662, 0.662, 0.662], [0.162, 0.662, 0.838],
-                            [0.838, 0.338, 0.162], [0.662, 0.838, 0.162],
-                            [0.338, 0.162, 0.838], [0.838, 0.162, 0.662],
-                            [0.162, 0.838, 0.338]],
-                "symbols": ['Fe'] * 4 + ['S'] * 8,
-                "pbc": [True, True, True]
-            }
-            return convert_structure(structure_data, "aiida")
-        raise ValueError(name)
-    return _get_structure
+    return get_test_structure
+
+
+@pytest.fixture(scope='function')
+def get_structure_and_symm():
+    return get_test_structure_and_symm
 
 
 @pytest.fixture(scope='function')
@@ -96,6 +76,22 @@ def get_cif():
         from aiida.plugins import DataFactory
         cif_data_cls = DataFactory('cif')
         if name == "pyrite":
-            return cif_data_cls(file=os.path.join(TEST_DIR, "cif_files", "pyrite.cif"))
+            return cif_data_cls(file=os.path.join(TEST_FILES, "cif", "pyrite.cif"))
         raise ValueError(name)
     return _get_cif
+
+
+@pytest.fixture(scope='function')
+def upload_basis_set_family():
+    """ upload the a basis set family"""
+    from aiida_crystal17.data.basis_set import BasisSetData
+
+    def _upload(folder_name="sto3g", group_name="sto3g", stop_if_existing=True):
+        BasisSetData.upload_basisset_family(
+            os.path.join(TEST_FILES, "basis_sets", folder_name),
+            group_name,
+            "minimal basis sets",
+            stop_if_existing=stop_if_existing,
+            extension=".basis")
+        return BasisSetData.get_basis_group_map(group_name)
+    return _upload
